move tests in own file
