+++ dot15d4/examples/parsing.rs
@@ -1,32 +1,32 @@
 use dot15d4::frame::*;
 
 fn main() {
-    let frame: [u8; 35] = [
+    let data: [u8; 35] = [
         0x40, 0xeb, 0xcd, 0xab, 0xff, 0xff, 0x01, 0x00, 0x01, 0x00, 0x01, 0x00, 0x01, 0x00, 0x00,
         0x3f, 0x11, 0x88, 0x06, 0x1a, 0x0e, 0x00, 0x00, 0x00, 0x00, 0x00, 0x01, 0x1c, 0x00, 0x01,
         0xc8, 0x00, 0x01, 0x1b, 0x00,
     ];
 
-    let frame = Frame::new(&frame[..]).unwrap();
+    let frame = Frame::new(&data[..]).unwrap();
     println!("{frame}");
-    let frame = FrameRepr::parse(&frame);
-    println!("{frame:#?}");
+    let parsed = FrameRepr::parse(&frame);
+    println!("{parsed:#?}");
 
-    let frame = [
+    let data = [
         0x02, 0x2e, 0x37, 0xcd, 0xab, 0x02, 0x00, 0x02, 0x00, 0x02, 0x00, 0x02, 0x00, 0x02, 0x0f,
         0xe1, 0x8f,
     ];
-    let frame = Frame::new(&frame[..]).unwrap();
+    let frame = Frame::new(&data[..]).unwrap();
     println!("{frame}");
-    let frame = FrameRepr::parse(&frame);
-    println!("{frame:#?}");
+    let parsed = FrameRepr::parse(&frame);
+    println!("{parsed:#?}");
 
-    let frame = [
+    let data = [
         0x41, 0xd8, 0x01, 0xcd, 0xab, 0xff, 0xff, 0xc7, 0xd9, 0xb5, 0x14, 0x00, 0x4b, 0x12, 0x00,
         0x2b, 0x00, 0x00, 0x00,
     ];
-    let frame = Frame::new(&frame[..]).unwrap();
+    let frame = Frame::new(&data[..]).unwrap();
     println!("{frame}");
-    let frame = FrameRepr::parse(&frame);
-    println!("{frame:#?}");
+    let parsed = FrameRepr::parse(&frame);
+    println!("{parsed:#?}");
 }
