+++ dot15d4/src/frame/mod.rs
@@ -166,6 +166,9 @@
 //! [`payload_information_elements`]: InformationElements::payload_information_elements
 //! [`nested_information_elements`]: PayloadInformationElement::nested_information_elements
 
+#[cfg(test)]
+mod tests;
+
 mod frame_control;
 pub use frame_control::FrameControl;
 pub use frame_control::FrameControlRepr;
@@ -346,17 +349,17 @@ impl<T: AsRef<[u8]> + AsMut<[u8]>> Frame<T> {
     pub fn set_addressing_fields(&mut self, addressing_fields: &AddressingFieldsRepr) {
         let start = 2 + (!self.frame_control().sequence_number_suppression() as usize);
 
-        let mut writer = AddressingFields::new(&mut self.buffer.as_mut()[start..]);
-        writer.write_fields(addressing_fields);
+        let mut w = AddressingFields::new(&mut self.buffer.as_mut()[start..]);
+        w.write_fields(addressing_fields);
     }
 
     /// Set the Auxiliary Security Header field values in the buffer, based on the given _.
-    pub fn set_aux_secu_header(&mut self) {
+    pub fn set_aux_sec_header(&mut self) {
         todo!();
     }
 
     /// Set the Information Elements field values in the buffer, based on the given _.
-    pub fn set_information_elements(&mut self) {
+    pub fn set_information_elements(&mut self, ie: &InformationElementsRepr) {
         todo!();
     }
 
@@ -434,6 +437,7 @@ pub struct FrameRepr<'p> {
 }
 
 impl<'r, 'f: 'r> FrameRepr<'f> {
+    /// Parse an IEEE 802.15.4 frame into a high-level representation.
     pub fn parse(frame: &'r Frame<&'f [u8]>) -> FrameRepr<'f> {
         Self {
             frame_control: FrameControlRepr::parse(frame.frame_control()),
@@ -463,209 +467,5 @@ impl<'r, 'f: 'r> FrameRepr<'f> {
             }),
             payload: frame.payload().unwrap_or(&[]),
         }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn ack_frame() {
-        let frame = [
-            0x02, 0x2e, 0x37, 0xcd, 0xab, 0x02, 0x00, 0x02, 0x00, 0x02, 0x00, 0x02, 0x00, 0x02,
-            0x0f, 0xe1, 0x8f,
-        ];
-
-        let frame = Frame::new(&frame).unwrap();
-
-        let fc = frame.frame_control();
-        assert_eq!(fc.frame_type(), FrameType::Ack);
-        assert!(!fc.security_enabled());
-        assert!(!fc.frame_pending());
-        assert!(!fc.ack_request());
-        assert!(!fc.pan_id_compression());
-        assert!(!fc.sequence_number_suppression());
-        assert!(fc.information_elements_present());
-        assert!(fc.dst_addressing_mode() == AddressingMode::Extended);
-        assert!(fc.frame_version() == FrameVersion::Ieee802154);
-        assert!(fc.src_addressing_mode() == AddressingMode::Absent);
-
-        assert!(frame.sequence_number() == Some(55));
-
-        let addressing = frame.addressing();
-        assert_eq!(addressing.dst_pan_id(&fc), Some(0xabcd));
-        assert_eq!(
-            addressing.dst_address(&fc),
-            Some(Address::Extended([
-                0x00, 0x02, 0x00, 0x02, 0x00, 0x02, 0x00, 0x02
-            ]))
-        );
-        assert_eq!(addressing.src_pan_id(&fc), None);
-        assert_eq!(addressing.src_address(&fc), Some(Address::Absent));
-
-        let ie = frame.information_elements().unwrap();
-        let mut headers = ie.header_information_elements();
-
-        let time_correction = headers.next().unwrap();
-        assert_eq!(
-            time_correction.element_id(),
-            HeaderElementId::TimeCorrection
-        );
-
-        let time_correction = TimeCorrection::new(time_correction.content());
-        assert_eq!(time_correction.len(), 2);
-        assert_eq!(
-            time_correction.time_correction(),
-            crate::time::Duration::from_us(-31)
-        );
-        assert!(time_correction.nack());
-    }
-
-    #[test]
-    fn data_frame() {
-        let frame = [
-            0x41, 0xd8, 0x01, 0xcd, 0xab, 0xff, 0xff, 0xc7, 0xd9, 0xb5, 0x14, 0x00, 0x4b, 0x12,
-            0x00, 0x2b, 0x00, 0x00, 0x00,
-        ];
-
-        let frame = Frame::new(&frame).unwrap();
-
-        let fc = frame.frame_control();
-        assert_eq!(fc.frame_type(), FrameType::Data);
-        assert!(!fc.security_enabled());
-        assert!(!fc.frame_pending());
-        assert!(!fc.ack_request());
-        assert!(fc.pan_id_compression());
-
-        assert!(fc.dst_addressing_mode() == AddressingMode::Short);
-        assert!(fc.frame_version() == FrameVersion::Ieee802154_2006);
-        assert!(fc.src_addressing_mode() == AddressingMode::Extended);
-
-        assert!(frame.payload().unwrap() == &[0x2b, 0x00, 0x00, 0x00][..]);
-    }
-
-    #[test]
-    fn enhanced_beacon() {
-        let frame: [u8; 35] = [
-            0x40, 0xeb, 0xcd, 0xab, 0xff, 0xff, 0x01, 0x00, 0x01, 0x00, 0x01, 0x00, 0x01, 0x00,
-            0x00, 0x3f, 0x11, 0x88, 0x06, 0x1a, 0x0e, 0x00, 0x00, 0x00, 0x00, 0x00, 0x01, 0x1c,
-            0x00, 0x01, 0xc8, 0x00, 0x01, 0x1b, 0x00,
-        ];
-
-        let frame = Frame::new(&frame).unwrap();
-        let fc = frame.frame_control();
-        assert_eq!(fc.frame_type(), FrameType::Beacon);
-        assert!(!fc.security_enabled());
-        assert!(!fc.frame_pending());
-        assert!(!fc.ack_request());
-        assert!(fc.pan_id_compression());
-        assert!(fc.sequence_number_suppression());
-        assert!(fc.information_elements_present());
-        assert_eq!(fc.dst_addressing_mode(), AddressingMode::Short);
-        assert_eq!(fc.src_addressing_mode(), AddressingMode::Extended);
-        assert_eq!(fc.frame_version(), FrameVersion::Ieee802154);
-
-        let addressing = frame.addressing();
-        assert_eq!(addressing.dst_pan_id(&fc), Some(0xabcd),);
-        assert_eq!(addressing.src_pan_id(&fc), None,);
-        assert_eq!(addressing.dst_address(&fc), Some(Address::BROADCAST));
-        assert_eq!(
-            addressing.src_address(&fc),
-            Some(Address::Extended([
-                0x00, 0x01, 0x00, 0x01, 0x00, 0x01, 0x00, 0x01
-            ]))
-        );
-        assert_eq!(addressing.len(&fc), 12);
-
-        let ie = frame.information_elements().unwrap();
-
-        let mut headers = ie.header_information_elements();
-        let terminator = headers.next().unwrap();
-        assert_eq!(terminator.element_id(), HeaderElementId::HeaderTermination1);
-        assert_eq!(terminator.len(), 0);
-
-        assert_eq!(headers.next(), None);
-
-        let mut payloads = ie.payload_information_elements();
-
-        let mlme = payloads.next().unwrap();
-        assert_eq!(mlme.group_id(), PayloadGroupId::Mlme);
-        assert_eq!(mlme.len() + 2, 19);
-        assert_eq!(payloads.next(), None);
-
-        let mut nested_iterator = NestedInformationElementsIterator::new(mlme.content());
-
-        let tsch_sync = nested_iterator.next().unwrap();
-        assert_eq!(
-            tsch_sync.sub_id(),
-            NestedSubId::Short(NestedSubIdShort::TschSynchronization)
-        );
-        assert_eq!(
-            TschSynchronization::new(tsch_sync.content()).absolute_slot_number(),
-            14
-        );
-        assert_eq!(
-            TschSynchronization::new(tsch_sync.content()).join_metric(),
-            0
-        );
-
-        let tsch_timeslot = nested_iterator.next().unwrap();
-        assert_eq!(
-            tsch_timeslot.sub_id(),
-            NestedSubId::Short(NestedSubIdShort::TschTimeslot)
-        );
-        assert_eq!(TschTimeslot::new(tsch_timeslot.content()).id(), 0);
-
-        let channel_hopping = nested_iterator.next().unwrap();
-        assert_eq!(
-            channel_hopping.sub_id(),
-            NestedSubId::Long(NestedSubIdLong::ChannelHopping)
-        );
-        assert_eq!(
-            ChannelHopping::new(channel_hopping.content()).hopping_sequence_id(),
-            0
-        );
-
-        let slotframe = nested_iterator.next().unwrap();
-        assert_eq!(
-            slotframe.sub_id(),
-            NestedSubId::Short(NestedSubIdShort::TschSlotframeAndLink)
-        );
-        assert_eq!(
-            TschSlotframeAndLink::new(slotframe.content()).number_of_slot_frames(),
-            0
-        );
-
-        assert_eq!(nested_iterator.next(), None);
-        assert!(frame.payload().is_none());
-    }
-
-    #[test]
-    fn write_buffer() {
-        let mut buffer = [0u8; 127];
-
-        let mut writer = Frame::new_unchecked(&mut buffer);
-        writer.set_frame_control(&FrameControlRepr {
-            frame_type: FrameType::Beacon,
-            security_enabled: false,
-            frame_pending: false,
-            ack_request: false,
-            pan_id_compression: true,
-            sequence_number_suppression: true,
-            information_elements_present: true,
-            dst_addressing_mode: AddressingMode::Short,
-            src_addressing_mode: AddressingMode::Extended,
-            frame_version: FrameVersion::Ieee802154,
-        });
-
-        writer.set_addressing_fields(&AddressingFieldsRepr {
-            dst_pan_id: None,
-            src_pan_id: None,
-            dst_address: None,
-            src_address: None,
-        });
-
-        assert_eq!(&buffer[..2], &[0x40, 0xeb]);
     }
 }
